requirements (1).txt
streamlit
tensorflow
numpy
pillow
opencv-python
matplotlib
scikit-learn
opencv-python